Refactor ChargeClient to use pydantic types
